requirements.txt
PyPDF2==3.0.1
reportlab==4.2.2
Pillow>=11.1.0
pdf2image==1.17.0
PyMuPDF
PyQt6